Create the history page components
Create all the necessary components for the history page, module, ctrl,
service, directive, tpl.
Create a directive to hold and reuse the history question list.
Create a json file with mocked data to be consumed by the history
service.
Add the new directories to the gruntfile.
Add the desired styles to the history question list.