Add new question fetching validation. *Add new get random track logic.
The track Data to practice is selected from the dashboard, once the user select this and navigate to practice, most of the request are based on the track data which is stored in memory. If the user refresh the page from practice section it will lose the track data. So, I added a new feature to to prevent this and allow the users continue practicing even if they refresh the page or change the product from the practice page. This new logic will request a track based on the group Id, select a random one. And finally show a set of questions based on the new random track info. (Still testing).