Work on the Immediate Fixes in the Bugs document
Add target="_blank" to the history page links.
Remove ‘question 1 of x’ on the top of the question.
Removed Whats New from the top navbar. Added a link "Return to
Dashboard" instead.
Fixed spacing for % of users text. Updated text to say "x% of users
chose this answer" as well.
Removed the question id from the history table, renamed Preview to
Question preview.
Whenever you click on a menu option, it is now set to active if it
doesn't navigate to another page.