Make the timeline broader, fix timeline scroll
Add negative margin to the history container, to make it look like it
takes the whole screen, add some paddings to the title.
Remove the watch in the setHeight directive, this is what causes the
scroll to go to the top when new elements are loaded, instead use the
resize event directly.