Make the timeline day's collapsable.
Add the logic to toggle a boolean flag in the directive to determine
whether a day is collapsed or not.
Create a service to cache the collapse states on which the days are.
Add an onCollapse callback in the history controller to handle the case
where all days are collapsed, when this happens the controller shoud
load more questions.